Rename jsondb module to textdb
