+++ src/legendmeta/textdb.py
